\name{CGHnormaliter-internal}
\title{Internal CGHnormaliter functions}
\alias{.readCghRaw}
\alias{.calculateMA}
\alias{.runCGHcall}
\alias{.extractNormals}
\alias{.localLowess}
\alias{.plotMA}

\description{Internal CGHnormaliter functions.}
\details{These are not to be called by the user.}
\keyword{internal}
