\name{CGHnormaliter-internal}
\title{Internal CGHnormaliter functions}
\alias{.cghRaw_read}
\alias{.cghRaw_ma}
\alias{.iterate_normalize_call}
\alias{.extract_normals}
\alias{.local_lowess}

\description{Internal CGHnormaliter functions.}
\details{These are not to be called by the user.}
\keyword{internal}
